Move style control group to separate class
